zam-plugins: Move ttl generarion from built time (quemu) to postinst
Qemu shipped by current oe-core crashes for some plugins and instead of turtle-
files core-dumps are installed - not helpful exacly.

Am aware that this causes another 1st boot delay but that is better than not
having zam-plugins. Once qemu is good enough to work for zam-plugins, we can
simply revert this commit.

Signed-off-by: Andreas Müller <schnitzeltony@gmail.com>